visao-computacional/requirements.txt
imutils==0.5.4
joblib==1.1.0
mahotas==1.4.13
numpy==1.23.1
opencv-python==4.6.0.66
pandas==1.4.3
Pillow==9.2.0
python-dateutil==2.8.2
pytz==2022.1
scikit-learn==1.1.2
scipy==1.9.0
six==1.16.0
sklearn==0.0
threadpoolctl==3.1.0
